refactor parts of ioc_factory into entity_factory
This is a result of the PR review by Gary. Some
other minor review points are also addressed.

Fix up the rest of the tests and remaining PR
comments will be in the next commit.
